Convert sdk tests to use config instead of test_client param
Related to #919
