+++ python_sdk/tests/integration/test_infrahub_client.py
@@ -1,29 +1,14 @@
-import json
-from typing import Any, Dict, Optional
-
-import httpx
 import pytest
-from fastapi.testclient import TestClient
 
 from infrahub.core import registry
 from infrahub.core.initialization import create_branch
 from infrahub.core.node import Node
 from infrahub_client import Config, InfrahubClient
 from infrahub_client.node import InfrahubNode
-from infrahub_client.types import HTTPMethod
+
+from .conftest import InfrahubTestClient
 
 # pylint: disable=unused-argument
-
-
-class InfrahubTestClient(TestClient):
-    async def async_request(
-        self, url: str, method: HTTPMethod, headers: Dict[str, Any], timeout: int, payload: Optional[Dict] = None
-    ) -> httpx.Response:
-        content = None
-        if payload:
-            content = str(json.dumps(payload)).encode("UTF-8")
-        with self as client:
-            return client.request(method=method.value, url=url, headers=headers, timeout=timeout, content=content)
 
 
 class TestInfrahubClient:
