+++ python_sdk/tests/integration/test_node.py
@@ -1,11 +1,12 @@
 import pytest
-from fastapi.testclient import TestClient
 
 from infrahub.core.manager import NodeManager
 from infrahub.core.node import Node
-from infrahub_client import InfrahubClient
+from infrahub_client import Config, InfrahubClient
 from infrahub_client.exceptions import NodeNotFound
 from infrahub_client.node import InfrahubNode
+
+from .conftest import InfrahubTestClient
 
 # pylint: disable=unused-argument
 
@@ -17,11 +18,12 @@ class TestInfrahubNode:
 
         from infrahub.server import app
 
-        return TestClient(app)
+        return InfrahubTestClient(app)
 
     @pytest.fixture
     async def client(self, test_client):
-        return await InfrahubClient.init(test_client=test_client)
+        config = Config(requester=test_client.async_request)
+        return await InfrahubClient.init(config=config)
 
     async def test_node_create(self, client: InfrahubClient, init_db_base, location_schema):
         data = {"name": {"value": "JFK1"}, "description": {"value": "JFK Airport"}, "type": {"value": "SITE"}}
